pyproject.toml
[project]
name = "chessdotkoriat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "python-chess>=1.999",
]
